+++ UIComponents.Generators/Generators/FormButtons/UICGeneratorButtonCreate.cs
@@ -15,18 +15,32 @@ public class UICGeneratorButtonCreate : UICGeneratorProperty
 
     public override async Task<IUICGeneratorResponse<IUIComponent>> GetResponseAsync(UICPropertyArgs args, IUIComponent? existingResult)
     {
-        if(args.Configuration.TryGetPermissionService(out var permissionService))
+        if (existingResult != null)
+            return GeneratorHelper.Next();
+
+        if(args.ClassObject != null && args.Configuration.TryGetPermissionService(out var permissionService))
         {
             if (!await permissionService!.CanCreateType(args.ClassObject.GetType()))
+            {
+                _logger.LogDebug("No Createbutton is created because there is no permission to create for type {0}", args.ClassObject.GetType().Name);
                 return GeneratorHelper.Success<IUIComponent>(null, false);
+            }
         }
-        var form = args.CallCollection.Components.Where(c => c is UICForm).OfType<UICForm>().FirstOrDefault();
-        if (form == null || form.TriggerSubmit() == null)
-            return GeneratorHelper.Next();
+        if(args.CallCollection.Caller != null)
+        {
+            var form = args.CallCollection.Components.Where(c => c is UICForm).OfType<UICForm>().FirstOrDefault();
+            if (form == null || form.TriggerSubmit() == null)
+            {
+                _logger.LogDebug("No Createbutton is created because no form was found");
+                return GeneratorHelper.Next();
+            }
+        }
+        
 
         var button = new UICButtonSave()
         {
             ButtonText = TranslationDefaults.ButtonCreate,
+            PrependButtonIcon = new UICIcon(IconDefaults.Create?.Icon??null)
         }.AddClass("btn-create");
 
         await Task.Delay(0);