+++ UIComponents.Generators/Generators/FormButtons/UICGeneratorButtonSave.cs
@@ -16,15 +16,28 @@ public class UICGeneratorButtonSave : UICGeneratorProperty
 
     public override async Task<IUICGeneratorResponse<IUIComponent>> GetResponseAsync(UICPropertyArgs args, IUIComponent? existingResult)
     {
-        if(args.Configuration.TryGetPermissionService(out var permissionService))
+        if (existingResult != null)
+            return GeneratorHelper.Next();
+        if(args.ClassObject != null && args.Configuration.TryGetPermissionService(out var permissionService))
         {
             if (!await permissionService.CanEditObject(args.ClassObject))
+            {
+                _logger.LogDebug("No SaveButton is created because there is no permission to edit this object {0} ({1})", args.ClassObject.ToString(), args.ClassObject.GetType().Name);
                 return GeneratorHelper.Success<IUIComponent>(null, false);
+            }
+                
         }
 
-        var form = args.CallCollection.Components.Where(c => c is UICForm).OfType<UICForm>().FirstOrDefault();
-        if (form == null || form.TriggerSubmit() == null)
-            return GeneratorHelper.Next();
+        if(args.CallCollection.Caller != null)
+        {
+            var form = args.CallCollection.Components.Where(c => c is UICForm).OfType<UICForm>().FirstOrDefault();
+            if (form == null || form.TriggerSubmit() == null)
+            {
+                _logger.LogDebug("No SaveButton is created because no form was found");
+                return GeneratorHelper.Next();
+            }
+        }
+        
 
         var button = new UICButtonSave()
         {